ws-server/requirements.txt
python-socketio==5.10.0
uvicorn[standard]==0.24.0
python-engineio==4.8.0
fastapi==0.104.1
asyncio-mqtt==0.16.1 